seperated Column from Board
