+++ src/Components/Tasks/Column.js
@@ -1,0 +1,42 @@
+import React from "react";
+import { getStatusFromColumnId } from "../../Pages/Board";
+import DisplayTask from "./DisplayTasks";
+import { Droppable, Draggable } from "react-beautiful-dnd";
+import "../../Pages/ProjPage.css";
+
+export default function Column(props) {
+  const { columnId, tasks } = props;
+  const columnTasks = tasks.filter(
+    (task) => task.status === getStatusFromColumnId(columnId)
+  );
+
+  return (
+    <div className="column" id={columnId}>
+      <h2>{columnId}</h2>
+      <Droppable droppableId={columnId}>
+        {(provided) => (
+          <div {...provided.droppableProps} ref={provided.innerRef}>
+            {columnTasks.map((task, index) => (
+              <Draggable
+                key={task.id}
+                draggableId={task.id.toString()}
+                index={index}
+              >
+                {(provided) => (
+                  <div
+                    ref={provided.innerRef}
+                    {...provided.draggableProps}
+                    {...provided.dragHandleProps}
+                  >
+                    <DisplayTask key={task.id} task={task} index={index} />
+                  </div>
+                )}
+              </Draggable>
+            ))}
+            {provided.placeholder}
+          </div>
+        )}
+      </Droppable>
+    </div>
+  );
+}
